# [Inductor][CPP] Optimize the epilogue for int8 GEMM Template