--- a/torch/_inductor/codegen/cpp_gemm_template.py
+++ b/torch/_inductor/codegen/cpp_gemm_template.py
@@ -318,6 +318,15 @@ GEMM_TEMPLATE = r"""
 }
 """
 
+def _is_int8_gemm(inputs):
+    return (
+        isinstance(inputs[0], ir.IRNode)
+        and inputs[0].get_dtype() in [torch.uint8, torch.int8]
+    ) or (
+        isinstance(inputs[0], torch.Tensor)
+        and inputs[0].dtype in [torch.uint8, torch.int8]
+    )
+
 
 def get_padded_n(n, block_n):
     return (n + block_n - 1) // block_n * block_n
@@ -717,6 +726,15 @@ class CppGemmTemplate(CppTemplate):
             dtype_B = self.input_nodes[1].get_dtype()
             num_byte_A = get_num_byte(dtype_A)
             num_byte_B = get_num_byte(dtype_B)
+
+            dtype_C = self.output_node.get_dtype()
+            num_byte_C = get_num_byte(dtype_C)
+
+            # print("---- num_byte_A is: {}".format(num_byte_A), flush=True)
+            # print("---- num_byte_B is: {}".format(num_byte_B), flush=True)
+            # print("---- num_byte_C is: {}".format(num_byte_C), flush=True)
+
+
             if dtype_A is torch.bfloat16 and dtype_B is torch.int8 and Kr != 1:
                 # We will cache dequantized weights (BF16) in L1D for AMX micro-kernel.
                 # In this case, the choice of the micro-kernel being used can't be decoupled from
@@ -754,6 +772,8 @@ class CppGemmTemplate(CppTemplate):
             else:
                 # Strategy 2: Kt is too large to hold A (Mc x Kt) in L2, we reuse
                 # A (Mc x Kc) in L2 by B (Kc x Nc). C (Mc x Nc) resides in L2.
+                # print("---- hit the strategy 2", flush=True)
+                breakpoint()
                 Mc_blocks = Mt_blocks
                 Nc_blocks = min(math.ceil(Mc_blocks * Mr / Nr), Nt_blocks)
                 Nc_bytes = Nc_blocks * Nr * 4  # assume C or acc is float32/int32
@@ -825,6 +845,16 @@ class CppGemmTemplate(CppTemplate):
         epilogue_creator: Optional[Callable[[ir.Buffer], ir.Pointwise]] = None,
         act_mapping: Optional[dict[int, ir.IRNode]] = None,
     ):
+        use_int8_fast_path = False
+        if _is_int8_gemm(input_nodes):
+            if (
+                isinstance(input_nodes[2], ir.TensorBox)
+                and isinstance(input_nodes[2].data.data, ir.ConstantBuffer)
+                and isinstance(input_nodes[4], ir.TensorBox)
+                and isinstance(input_nodes[4].data.data, ir.ConstantBuffer)
+            ):
+                use_int8_fast_path = True
+
         if input_indices is None:
             input_indices = list(range(len(input_nodes)))
         only_one_input = (
@@ -937,7 +967,7 @@ class CppGemmTemplate(CppTemplate):
             if only_one_input and isinstance(new_inputs[0], torch.Tensor):
                 return new_inputs[1:], new_layout
             return cls.prep_weight(
-                new_inputs, new_layout, micro_gemm, pre_block_weights
+                new_inputs, new_layout, micro_gemm, pre_block_weights, use_int8_fast_path
             )
 
         def postprocessor(output):
@@ -1002,6 +1032,7 @@ class CppGemmTemplate(CppTemplate):
         layout: ir.Layout,
         micro_gemm: CppMicroGemm,
         should_block_weight: bool,
+        use_int8_fast_path: bool = False,
     ):
         """
         NOTE Weight prep consists of 2 separate steps:
@@ -1046,26 +1077,46 @@ class CppGemmTemplate(CppTemplate):
             # Require W layout to be fixed & contiguous, happens inplace.
             ir.ExternKernel.require_contiguous(W)
 
-        def _is_int8_gemm(inputs):
-            return (
-                isinstance(inputs[0], ir.IRNode)
-                and inputs[0].get_dtype() in [torch.uint8, torch.int8]
-            ) or (
-                isinstance(inputs[0], torch.Tensor)
-                and inputs[0].dtype in [torch.uint8, torch.int8]
-            )
 
         if _is_int8_gemm(new_inputs):
             BCompensate = None
-            if isinstance(W, ir.IRNode):
-                BCompensate = V.graph.add_tensor_constant(
-                    V.graph.constants[W.get_name() + "_BMatrixCompens"],
-                    W.get_name() + "_BMatrixCompens",
-                )
+            if use_int8_fast_path:
+                x_scale = new_inputs[2]
+                w_scale = new_inputs[4]
+                x_zp = new_inputs[3]
+                if isinstance(W, ir.IRNode):
+                    BCompensate = V.graph.add_tensor_constant(
+                        V.graph.constants[W.get_name() + "_BMatrixCompens"],
+                        W.get_name() + "_BMatrixCompens",
+                    )
+                else:
+                    # Use the original W, not the blocked_w in new_inputs[1] to calculate BCompensate
+                    BCompensate = torch.sum(W.to_dense().to(torch.float), dim=0)  # type: ignore[assignment]
+                    BCompensate = BCompensate * x_scale * w_scale * x_zp
+                
+                new_inputs.append(BCompensate)
+
+                x_w_scale = None
+                if isinstance(W, ir.IRNode):
+                    x_w_scale = V.graph.add_tensor_constant(
+                        V.graph.constants[W.get_name() + "_x_w_compens"],
+                        W.get_name() + "_x_w_compens",
+                    )
+                else:
+                    assert isinstance(x_scale, torch.Tensor)
+                    assert isinstance(w_scale, torch.Tensor)
+                    x_w_scale = x_scale * w_scale
+                new_inputs.append(x_w_scale)
             else:
-                # Use the original W, not the blocked_w in new_inputs[1] to calculate BCompensate
-                BCompensate = torch.sum(W.to_dense().to(torch.float), dim=0)  # type: ignore[assignment]
-            new_inputs.append(BCompensate)
+                if isinstance(W, ir.IRNode):
+                    BCompensate = V.graph.add_tensor_constant(
+                        V.graph.constants[W.get_name() + "_BMatrixCompens"],
+                        W.get_name() + "_BMatrixCompens",
+                    )
+                else:
+                    # Use the original W, not the blocked_w in new_inputs[1] to calculate BCompensate
+                    BCompensate = torch.sum(W.to_dense().to(torch.float), dim=0)  # type: ignore[assignment]
+                new_inputs.append(BCompensate)
         return new_inputs, layout
 
     @staticmethod